+++ example/src/app.js
@@ -4,12 +4,51 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Geosuggest from '../../src/Geosuggest';
 
-var App = React.createClass({ // eslint-disable-line
+class App extends React.Component {
+  /**
+   * When the input receives focus
+   */
+  onFocus() {
+    console.log('onFocus'); // eslint-disable-line
+  }
+
+  /**
+   * When the input loses focus
+   * @param {String} value The user input
+   */
+  onBlur(value) {
+    console.log('onBlur', value); // eslint-disable-line
+  }
+
+  /**
+   * When the input got changed
+   * @param {String} value The new value
+   */
+  onChange(value) {
+    console.log('input changes to :' + value); // eslint-disable-line
+  }
+
+  /**
+   * When a suggest got selected
+   * @param  {Object} suggest The suggest
+   */
+  onSuggestSelect(suggest) {
+    console.log(suggest); // eslint-disable-line
+  }
+
+  /**
+   * When there are no suggest results
+   * @param {String} userInput The user input
+   */
+  onSuggestNoResults(userInput) {
+    console.log('onSuggestNoResults for :' + userInput); // eslint-disable-line
+  }
+
   /**
    * Render the example app
    * @return {Function} React render function
    */
-  render: function() {
+  render() {
     var fixtures = [
       {label: 'New York', location: {lat: 40.7033127, lng: -73.979681}},
       {label: 'Rio', location: {lat: -22.066452, lng: -42.9232368}},
@@ -29,46 +68,7 @@ var App = React.createClass({ // eslint-disable-line
           radius="20" />
       </div>
     );
-  },
-
-  /**
-   * When the input receives focus
-   */
-  onFocus: function() {
-    console.log('onFocus'); // eslint-disable-line
-  },
-
-  /**
-   * When the input loses focus
-   * @param {String} value The user input
-   */
-  onBlur: function(value) {
-    console.log('onBlur', value); // eslint-disable-line
-  },
-
-  /**
-   * When the input got changed
-   * @param {String} value The new value
-   */
-  onChange: function(value) {
-    console.log('input changes to :' + value); // eslint-disable-line
-  },
-
-  /**
-   * When a suggest got selected
-   * @param  {Object} suggest The suggest
-   */
-  onSuggestSelect: function(suggest) {
-    console.log(suggest); // eslint-disable-line
-  },
-
-  /**
-   * When there are no suggest results
-   * @param {String} userInput The user input
-   */
-  onSuggestNoResults: function(userInput) {
-    console.log('onSuggestNoResults for :' + userInput); // eslint-disable-line
   }
-});
+}
 
 ReactDOM.render(<App />, document.getElementById('app')); // eslint-disable-line