+++ src/suggest-list.jsx
@@ -1,5 +1,5 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
-import shallowCompare from 'react/lib/shallowCompare';
+import shallowCompare from 'react-addons-shallow-compare';
 import classnames from 'classnames';
 import SuggestItem from './suggest-item';
 
@@ -55,10 +55,13 @@ export default class SuggestList extends React.Component {
     return <ul className={classes} style={this.props.style}>
       {this.props.suggests.map(suggest => {
         const isActive = this.props.activeSuggest &&
-          suggest.placeId === this.props.activeSuggest.placeId;
+            suggest.placeId === this.props.activeSuggest.placeId,
+          key = suggest.key || suggest.placeId;
 
-        return <SuggestItem key={suggest.placeId}
+        return <SuggestItem key={key}
           className={suggest.className}
+          userInput={this.props.userInput}
+          isHighlightMatch={this.props.isHighlightMatch}
           suggest={suggest}
           style={this.props.suggestItemStyle}
           suggestItemClassName={this.props.suggestItemClassName}
@@ -66,7 +69,8 @@ export default class SuggestList extends React.Component {
           activeClassname={this.props.suggestItemActiveClassName}
           onMouseDown={this.props.onSuggestMouseDown}
           onMouseOut={this.props.onSuggestMouseOut}
-          onSelect={this.props.onSuggestSelect} />;
+          onSelect={this.props.onSuggestSelect}
+          renderSuggestItem={this.props.renderSuggestItem}/>;
       })}
     </ul>;
   }